chore: Refactored import service to use type parameterized generic interface impls (#28245)
