+++ app/server/appsmith-server/src/main/java/com/appsmith/server/actioncollections/base/ActionCollectionServiceCEImpl.java
@@ -8,18 +8,14 @@ import com.appsmith.server.acl.PolicyGenerator;
 import com.appsmith.server.constants.FieldName;
 import com.appsmith.server.domains.Action;
 import com.appsmith.server.domains.ActionCollection;
-import com.appsmith.server.domains.Application;
 import com.appsmith.server.domains.NewPage;
 import com.appsmith.server.domains.Page;
 import com.appsmith.server.dtos.ActionCollectionDTO;
 import com.appsmith.server.dtos.ActionCollectionViewDTO;
-import com.appsmith.server.dtos.ce.ImportActionCollectionResultDTO;
-import com.appsmith.server.dtos.ce.ImportActionResultDTO;
 import com.appsmith.server.exceptions.AppsmithError;
 import com.appsmith.server.exceptions.AppsmithException;
 import com.appsmith.server.helpers.DefaultResourcesUtils;
 import com.appsmith.server.helpers.ResponseUtils;
-import com.appsmith.server.helpers.ce.ImportApplicationPermissionProvider;
 import com.appsmith.server.newactions.base.NewActionService;
 import com.appsmith.server.repositories.ActionCollectionRepository;
 import com.appsmith.server.services.AnalyticsService;
@@ -45,7 +41,6 @@ import reactor.core.scheduler.Scheduler;
 
 import java.time.Instant;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -54,7 +49,6 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static com.appsmith.external.helpers.AppsmithBeanUtils.copyNestedNonNullProperties;
 import static com.appsmith.external.helpers.AppsmithBeanUtils.copyNewFieldValuesIntoOldObject;
 import static java.lang.Boolean.TRUE;
 
@@ -625,220 +619,5 @@ public class ActionCollectionServiceCEImpl extends BaseService<ActionCollectionR
         actionCollection.setDeleted(branchedActionCollection.getDeleted());
         // Set policies from existing branch object
         actionCollection.setPolicies(branchedActionCollection.getPolicies());
-    }
-
-    private NewPage updatePageInActionCollection(ActionCollectionDTO collectionDTO, Map<String, NewPage> pageNameMap) {
-        NewPage parentPage = pageNameMap.get(collectionDTO.getPageId());
-        if (parentPage == null) {
-            return null;
-        }
-        collectionDTO.setPageId(parentPage.getId());
-
-        // Update defaultResources in actionCollectionDTO
-        DefaultResources defaultResources = new DefaultResources();
-        defaultResources.setPageId(parentPage.getDefaultResources().getPageId());
-        collectionDTO.setDefaultResources(defaultResources);
-
-        return parentPage;
-    }
-
-    /**
-     * Method to
-     * - save imported actionCollections with updated policies
-     * - update default resource ids along with branch-name if the application is connected to git
-     *
-     * @param importedActionCollectionList          action list extracted from the imported JSON file
-     * @param application                   imported and saved application in DB
-     * @param branchName                            branch to which the actions needs to be saved if the application is connected to git
-     * @param pageNameMap                           map of page name to saved page in DB
-     * @param pluginMap                             map of plugin name to saved plugin id in DB
-     * @param permissionProvider
-     * @return tuple of imported actionCollectionId and saved actionCollection in DB
-     */
-    @Override
-    public Mono<ImportActionCollectionResultDTO> importActionCollections(
-            ImportActionResultDTO importActionResultDTO,
-            Application application,
-            String branchName,
-            List<ActionCollection> importedActionCollectionList,
-            Map<String, String> pluginMap,
-            Map<String, NewPage> pageNameMap,
-            ImportApplicationPermissionProvider permissionProvider) {
-
-        /* Mono.just(application) is created to avoid the eagerly fetching of existing actionCollections
-         * during the pipeline construction. It should be fetched only when the pipeline is subscribed/executed.
-         */
-        return Mono.just(application)
-                .flatMap(importedApplication -> {
-                    ImportActionCollectionResultDTO resultDTO = new ImportActionCollectionResultDTO();
-                    final String workspaceId = importedApplication.getWorkspaceId();
-
-                    // Map of gitSyncId to actionCollection of the existing records in DB
-                    Mono<Map<String, ActionCollection>> actionCollectionsInCurrentAppMono = repository
-                            .findByApplicationId(importedApplication.getId())
-                            .filter(collection -> collection.getGitSyncId() != null)
-                            .collectMap(ActionCollection::getGitSyncId);
-
-                    Mono<Map<String, ActionCollection>> actionCollectionsInBranchesMono;
-                    if (importedApplication.getGitApplicationMetadata() != null) {
-                        final String defaultApplicationId =
-                                importedApplication.getGitApplicationMetadata().getDefaultApplicationId();
-                        actionCollectionsInBranchesMono = repository
-                                .findByDefaultApplicationId(defaultApplicationId, Optional.empty())
-                                .filter(actionCollection -> actionCollection.getGitSyncId() != null)
-                                .collectMap(ActionCollection::getGitSyncId);
-                    } else {
-                        actionCollectionsInBranchesMono = Mono.just(Collections.emptyMap());
-                    }
-
-                    return Mono.zip(actionCollectionsInCurrentAppMono, actionCollectionsInBranchesMono)
-                            .flatMap(objects -> {
-                                Map<String, ActionCollection> actionsCollectionsInCurrentApp = objects.getT1();
-                                Map<String, ActionCollection> actionsCollectionsInBranches = objects.getT2();
-
-                                // set the existing action collections in the result DTO, this will be required in next
-                                // phases
-                                resultDTO.setExistingActionCollections(actionsCollectionsInCurrentApp.values());
-
-                                List<ActionCollection> newActionCollections = new ArrayList<>();
-                                List<ActionCollection> existingActionCollections = new ArrayList<>();
-
-                                for (ActionCollection actionCollection : importedActionCollectionList) {
-                                    if (actionCollection.getUnpublishedCollection() == null
-                                            || StringUtils.isEmpty(actionCollection
-                                                    .getUnpublishedCollection()
-                                                    .getPageId())) {
-                                        continue; // invalid action collection, skip it
-                                    }
-                                    final String idFromJsonFile = actionCollection.getId();
-                                    NewPage parentPage = new NewPage();
-                                    final ActionCollectionDTO unpublishedCollection =
-                                            actionCollection.getUnpublishedCollection();
-                                    final ActionCollectionDTO publishedCollection =
-                                            actionCollection.getPublishedCollection();
-
-                                    // If pageId is missing in the actionCollectionDTO create a fallback pageId
-                                    final String fallbackParentPageId = unpublishedCollection.getPageId();
-
-                                    if (unpublishedCollection.getName() != null) {
-                                        unpublishedCollection.setDefaultToBranchedActionIdsMap(importActionResultDTO
-                                                .getUnpublishedCollectionIdToActionIdsMap()
-                                                .get(idFromJsonFile));
-                                        unpublishedCollection.setPluginId(
-                                                pluginMap.get(unpublishedCollection.getPluginId()));
-                                        parentPage = updatePageInActionCollection(unpublishedCollection, pageNameMap);
-                                    }
-
-                                    if (publishedCollection != null && publishedCollection.getName() != null) {
-                                        publishedCollection.setDefaultToBranchedActionIdsMap(importActionResultDTO
-                                                .getPublishedCollectionIdToActionIdsMap()
-                                                .get(idFromJsonFile));
-                                        publishedCollection.setPluginId(
-                                                pluginMap.get(publishedCollection.getPluginId()));
-                                        if (StringUtils.isEmpty(publishedCollection.getPageId())) {
-                                            publishedCollection.setPageId(fallbackParentPageId);
-                                        }
-                                        NewPage publishedCollectionPage =
-                                                updatePageInActionCollection(publishedCollection, pageNameMap);
-                                        parentPage = parentPage == null ? publishedCollectionPage : parentPage;
-                                    }
-
-                                    actionCollection.makePristine();
-                                    actionCollection.setWorkspaceId(workspaceId);
-                                    actionCollection.setApplicationId(importedApplication.getId());
-
-                                    // Check if the action has gitSyncId and if it's already in DB
-                                    if (actionCollection.getGitSyncId() != null
-                                            && actionsCollectionsInCurrentApp.containsKey(
-                                                    actionCollection.getGitSyncId())) {
-
-                                        // Since the resource is already present in DB, just update resource
-                                        ActionCollection existingActionCollection =
-                                                actionsCollectionsInCurrentApp.get(actionCollection.getGitSyncId());
-
-                                        Set<Policy> existingPolicy = existingActionCollection.getPolicies();
-                                        copyNestedNonNullProperties(actionCollection, existingActionCollection);
-                                        // Update branchName
-                                        existingActionCollection
-                                                .getDefaultResources()
-                                                .setBranchName(branchName);
-                                        // Recover the deleted state present in DB from imported actionCollection
-                                        existingActionCollection
-                                                .getUnpublishedCollection()
-                                                .setDeletedAt(actionCollection
-                                                        .getUnpublishedCollection()
-                                                        .getDeletedAt());
-                                        existingActionCollection.setDeletedAt(actionCollection.getDeletedAt());
-                                        existingActionCollection.setDeleted(actionCollection.getDeleted());
-                                        existingActionCollection.setPolicies(existingPolicy);
-
-                                        existingActionCollection.updateForBulkWriteOperation();
-                                        existingActionCollections.add(existingActionCollection);
-                                        resultDTO.getSavedActionCollectionIds().add(existingActionCollection.getId());
-                                        resultDTO
-                                                .getSavedActionCollectionMap()
-                                                .put(idFromJsonFile, existingActionCollection);
-                                    } else {
-                                        if (!permissionProvider.canCreateAction(parentPage)) {
-                                            throw new AppsmithException(
-                                                    AppsmithError.ACL_NO_RESOURCE_FOUND,
-                                                    FieldName.PAGE,
-                                                    parentPage.getId());
-                                        }
-
-                                        if (importedApplication.getGitApplicationMetadata() != null) {
-                                            final String defaultApplicationId = importedApplication
-                                                    .getGitApplicationMetadata()
-                                                    .getDefaultApplicationId();
-                                            if (actionsCollectionsInBranches.containsKey(
-                                                    actionCollection.getGitSyncId())) {
-                                                ActionCollection branchedActionCollection =
-                                                        actionsCollectionsInBranches.get(
-                                                                actionCollection.getGitSyncId());
-                                                populateDefaultResources(
-                                                        actionCollection, branchedActionCollection, branchName);
-                                            } else {
-                                                DefaultResources defaultResources = new DefaultResources();
-                                                defaultResources.setApplicationId(defaultApplicationId);
-                                                defaultResources.setBranchName(branchName);
-                                                actionCollection.setDefaultResources(defaultResources);
-                                            }
-                                        }
-
-                                        // this will generate the id and other auto generated fields e.g. createdAt
-                                        actionCollection.updateForBulkWriteOperation();
-                                        generateAndSetPolicies(parentPage, actionCollection);
-
-                                        // create or update default resources for the action
-                                        // values already set to defaultResources are kept unchanged
-                                        DefaultResourcesUtils.createDefaultIdsOrUpdateWithGivenResourceIds(
-                                                actionCollection, branchName);
-
-                                        // generate gitSyncId if it's not present
-                                        if (actionCollection.getGitSyncId() == null) {
-                                            actionCollection.setGitSyncId(
-                                                    actionCollection.getApplicationId() + "_" + new ObjectId());
-                                        }
-
-                                        // it's new actionCollection
-                                        newActionCollections.add(actionCollection);
-                                        resultDTO.getSavedActionCollectionIds().add(actionCollection.getId());
-                                        resultDTO.getSavedActionCollectionMap().put(idFromJsonFile, actionCollection);
-                                    }
-                                }
-                                log.info(
-                                        "Saving action collections in bulk. New: {}, Updated: {}",
-                                        newActionCollections.size(),
-                                        existingActionCollections.size());
-                                return repository
-                                        .bulkInsert(newActionCollections)
-                                        .then(repository.bulkUpdate(existingActionCollections))
-                                        .thenReturn(resultDTO);
-                            });
-                })
-                .onErrorResume(e -> {
-                    log.error("Error saving action collections", e);
-                    return Mono.error(e);
-                });
     }
 }
