+++ app/server/appsmith-server/src/main/java/com/appsmith/server/actioncollections/exports/ActionCollectionExportableServiceImpl.java
@@ -1,4 +1,4 @@
-package com.appsmith.server.actioncollections.export;
+package com.appsmith.server.actioncollections.exports;
 
 import com.appsmith.server.actioncollections.base.ActionCollectionService;
 import com.appsmith.server.domains.ActionCollection;
