+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/MappedImportableResourcesDTO.java
@@ -1,0 +1,5 @@
+package com.appsmith.server.dtos;
+
+import com.appsmith.server.dtos.ce.MappedImportableResourcesCE_DTO;
+
+public class MappedImportableResourcesDTO extends MappedImportableResourcesCE_DTO {}
