+++ app/server/appsmith-server/src/main/java/com/appsmith/server/imports/importable/ImportableServiceCE.java
@@ -1,5 +1,28 @@
 package com.appsmith.server.imports.importable;
 
 import com.appsmith.external.models.BaseDomain;
+import com.appsmith.server.domains.Application;
+import com.appsmith.server.domains.Workspace;
+import com.appsmith.server.dtos.ApplicationJson;
+import com.appsmith.server.dtos.ImportingMetaDTO;
+import com.appsmith.server.dtos.MappedImportableResourcesDTO;
+import reactor.core.publisher.Mono;
 
-public interface ImportableServiceCE<T extends BaseDomain> {}
+import java.util.List;
+
+public interface ImportableServiceCE<T extends BaseDomain> {
+
+    Mono<List<T>> importEntities(
+            ImportingMetaDTO importingMetaDTO,
+            MappedImportableResourcesDTO mappedImportableResourcesDTO,
+            Mono<Workspace> workspaceMono,
+            Mono<Application> applicationMono,
+            ApplicationJson applicationJson);
+
+    default Mono<Void> updateImportedEntities(
+            Application application,
+            ImportingMetaDTO importingMetaDTO,
+            MappedImportableResourcesDTO mappedImportableResourcesDTO) {
+        return null;
+    }
+}
