+++ app/server/appsmith-server/src/main/java/com/appsmith/server/imports/internal/ImportApplicationServiceCEImpl.java
@@ -2,23 +2,9 @@ package com.appsmith.server.imports.internal;
 
 import com.appsmith.external.constants.AnalyticsEvents;
 import com.appsmith.external.helpers.Stopwatch;
-import com.appsmith.external.models.AuthenticationResponse;
-import com.appsmith.external.models.BaseDomain;
-import com.appsmith.external.models.BasicAuth;
-import com.appsmith.external.models.BearerTokenAuth;
-import com.appsmith.external.models.DBAuth;
 import com.appsmith.external.models.Datasource;
-import com.appsmith.external.models.DatasourceConfiguration;
-import com.appsmith.external.models.DatasourceStorage;
 import com.appsmith.external.models.DatasourceStorageDTO;
-import com.appsmith.external.models.DecryptedSensitiveFields;
-import com.appsmith.external.models.DefaultResources;
-import com.appsmith.external.models.OAuth2;
-import com.appsmith.external.models.Policy;
-import com.appsmith.server.actioncollections.base.ActionCollectionService;
-import com.appsmith.server.constants.ApplicationConstants;
 import com.appsmith.server.constants.FieldName;
-import com.appsmith.server.constants.ResourceModes;
 import com.appsmith.server.datasources.base.DatasourceService;
 import com.appsmith.server.domains.ActionCollection;
 import com.appsmith.server.domains.Application;
@@ -26,31 +12,26 @@ import com.appsmith.server.domains.ApplicationPage;
 import com.appsmith.server.domains.CustomJSLib;
 import com.appsmith.server.domains.NewAction;
 import com.appsmith.server.domains.NewPage;
+import com.appsmith.server.domains.Plugin;
+import com.appsmith.server.domains.Theme;
 import com.appsmith.server.domains.User;
 import com.appsmith.server.domains.Workspace;
 import com.appsmith.server.dtos.ApplicationImportDTO;
 import com.appsmith.server.dtos.ApplicationJson;
-import com.appsmith.server.dtos.CustomJSLibApplicationDTO;
-import com.appsmith.server.dtos.ce.ImportActionCollectionResultDTO;
-import com.appsmith.server.dtos.ce.ImportActionResultDTO;
-import com.appsmith.server.dtos.ce.ImportedActionAndCollectionMapsDTO;
+import com.appsmith.server.dtos.ImportingMetaDTO;
+import com.appsmith.server.dtos.MappedImportableResourcesDTO;
 import com.appsmith.server.exceptions.AppsmithError;
 import com.appsmith.server.exceptions.AppsmithException;
-import com.appsmith.server.helpers.DefaultResourcesUtils;
 import com.appsmith.server.helpers.ImportExportUtils;
-import com.appsmith.server.helpers.TextUtils;
 import com.appsmith.server.helpers.ce.ImportApplicationPermissionProvider;
-import com.appsmith.server.jslibs.base.CustomJSLibService;
+import com.appsmith.server.imports.importable.ImportableService;
 import com.appsmith.server.migrations.ApplicationVersion;
 import com.appsmith.server.migrations.JsonSchemaMigration;
 import com.appsmith.server.newactions.base.NewActionService;
-import com.appsmith.server.newpages.base.NewPageService;
 import com.appsmith.server.repositories.PermissionGroupRepository;
-import com.appsmith.server.repositories.PluginRepository;
 import com.appsmith.server.services.AnalyticsService;
 import com.appsmith.server.services.ApplicationPageService;
 import com.appsmith.server.services.ApplicationService;
-import com.appsmith.server.services.SequenceService;
 import com.appsmith.server.services.SessionUserService;
 import com.appsmith.server.services.WorkspaceService;
 import com.appsmith.server.solutions.ActionPermission;
@@ -58,39 +39,28 @@ import com.appsmith.server.solutions.ApplicationPermission;
 import com.appsmith.server.solutions.DatasourcePermission;
 import com.appsmith.server.solutions.PagePermission;
 import com.appsmith.server.solutions.WorkspacePermission;
-import com.appsmith.server.themes.base.ThemeService;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.bson.types.ObjectId;
 import org.springframework.core.io.buffer.DataBufferUtils;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.http.MediaType;
 import org.springframework.http.codec.multipart.Part;
 import org.springframework.transaction.reactive.TransactionalOperator;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import reactor.util.function.Tuple2;
-import reactor.util.function.Tuples;
 
 import java.lang.reflect.Type;
-import java.time.Instant;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static com.appsmith.external.helpers.AppsmithBeanUtils.copyNestedNonNullProperties;
-import static com.appsmith.server.constants.ResourceModes.EDIT;
-import static com.appsmith.server.constants.ResourceModes.VIEW;
 import static com.appsmith.server.helpers.ImportExportUtils.setPropertiesToExistingApplication;
 import static com.appsmith.server.helpers.ImportExportUtils.setPublishedApplicationProperties;
 
@@ -98,19 +68,15 @@ import static com.appsmith.server.helpers.ImportExportUtils.setPublishedApplicat
 @RequiredArgsConstructor
 public class ImportApplicationServiceCEImpl implements ImportApplicationServiceCE {
 
+    private static final Set<MediaType> ALLOWED_CONTENT_TYPES = Set.of(MediaType.APPLICATION_JSON);
+    private static final String INVALID_JSON_FILE = "invalid json file";
     private final DatasourceService datasourceService;
     private final SessionUserService sessionUserService;
-    private final PluginRepository pluginRepository;
     private final WorkspaceService workspaceService;
     private final ApplicationService applicationService;
-    private final NewPageService newPageService;
     private final ApplicationPageService applicationPageService;
     private final NewActionService newActionService;
-    private final SequenceService sequenceService;
-    private final ActionCollectionService actionCollectionService;
-    private final ThemeService themeService;
     private final AnalyticsService analyticsService;
-    private final CustomJSLibService customJSLibService;
     private final DatasourcePermission datasourcePermission;
     private final WorkspacePermission workspacePermission;
     private final ApplicationPermission applicationPermission;
@@ -119,9 +85,13 @@ public class ImportApplicationServiceCEImpl implements ImportApplicationServiceC
     private final Gson gson;
     private final TransactionalOperator transactionalOperator;
     private final PermissionGroupRepository permissionGroupRepository;
-
-    private static final Set<MediaType> ALLOWED_CONTENT_TYPES = Set.of(MediaType.APPLICATION_JSON);
-    private static final String INVALID_JSON_FILE = "invalid json file";
+    private final ImportableService<Plugin> pluginImportableService;
+    private final ImportableService<Theme> themeImportableService;
+    private final ImportableService<NewPage> newPageImportableService;
+    private final ImportableService<CustomJSLib> customJSLibImportableService;
+    private final ImportableService<Datasource> datasourceImportableService;
+    private final ImportableService<NewAction> newActionImportableService;
+    private final ImportableService<ActionCollection> actionCollectionImportableService;
 
     @Override
     public Mono<ApplicationImportDTO> extractFileAndSaveApplication(String workspaceId, Part filePart) {
@@ -380,146 +350,11 @@ public class ImportApplicationServiceCEImpl implements ImportApplicationServiceC
         return errorField;
     }
 
-    private Mono<Map<String, String>> importDatasources(
-            ApplicationJson importedDoc,
-            Mono<List<Datasource>> existingDatasourceMono,
-            Flux<Datasource> existingDatasourcesFlux,
-            Workspace workspace,
-            Mono<Map<String, String>> pluginMapMno,
-            ImportApplicationPermissionProvider permissionProvider) {
-        return Mono.zip(
-                        existingDatasourceMono,
-                        workspaceService.getDefaultEnvironmentId(workspace.getId(), null),
-                        pluginMapMno)
-                .flatMapMany(objects -> {
-                    List<Datasource> existingDatasources = objects.getT1();
-                    String environmentId = objects.getT2();
-                    Map<String, String> pluginMap = objects.getT3();
-                    List<DatasourceStorage> importedDatasourceList = importedDoc.getDatasourceList();
-                    if (CollectionUtils.isEmpty(importedDatasourceList)) {
-                        return Flux.empty();
-                    }
-                    Map<String, Datasource> savedDatasourcesGitIdToDatasourceMap = new HashMap<>();
-                    Map<String, Datasource> savedDatasourcesNameDatasourceMap = new HashMap<>();
-
-                    existingDatasources.stream()
-                            .filter(datasource -> datasource.getGitSyncId() != null)
-                            .forEach(datasource -> {
-                                savedDatasourcesGitIdToDatasourceMap.put(datasource.getGitSyncId(), datasource);
-                                savedDatasourcesNameDatasourceMap.put(datasource.getName(), datasource);
-                            });
-
-                    // Check if the destination org have all the required plugins installed
-                    for (DatasourceStorage datasource : importedDatasourceList) {
-                        if (StringUtils.isEmpty(pluginMap.get(datasource.getPluginId()))) {
-                            log.error(
-                                    "Unable to find the plugin: {}, available plugins are: {}",
-                                    datasource.getPluginId(),
-                                    pluginMap.keySet());
-                            return Flux.error(new AppsmithException(
-                                    AppsmithError.UNKNOWN_PLUGIN_REFERENCE, datasource.getPluginId()));
-                        }
-                    }
-
-                    return Flux.fromIterable(importedDatasourceList)
-                            // Check for duplicate datasources to avoid duplicates in target workspace
-                            .flatMap(datasourceStorage -> {
-                                final String importedDatasourceName = datasourceStorage.getName();
-                                // try to find whether there is an existing datasource with same gitSyncId
-                                Datasource existingDatasource = null;
-                                if (datasourceStorage.getGitSyncId() != null
-                                        && savedDatasourcesGitIdToDatasourceMap.containsKey(
-                                                datasourceStorage.getGitSyncId())) {
-                                    Datasource dsWithSameGitsyncId = savedDatasourcesGitIdToDatasourceMap.get(
-                                            datasourceStorage.getGitSyncId()); // found a match
-                                    // we'll be renaming the matchingDatasource with targetDatasourceName
-                                    String targetDatasourceName = datasourceStorage.getName();
-
-                                    // check whether there are any other datasource with the same targetDatasourceName
-                                    if (savedDatasourcesNameDatasourceMap.containsKey(targetDatasourceName)) {
-                                        // found with same name, check if it's matchingDatasource or not
-                                        Datasource dsWithSameName =
-                                                savedDatasourcesNameDatasourceMap.get(targetDatasourceName);
-                                        if (dsWithSameName.getId().equals(dsWithSameGitsyncId.getId())) {
-                                            // same DS, we can rename safely
-                                            existingDatasource = dsWithSameGitsyncId;
-                                        } // otherwise existingDatasource will be null
-                                    } else { // no DS found with the targetDatasourceName, we can rename safely
-                                        existingDatasource = dsWithSameGitsyncId;
-                                    }
-                                }
-                                // Check if the datasource has gitSyncId and if it's already in DB
-                                if (existingDatasource != null) {
-                                    // Since the resource is already present in DB, just update resource
-                                    if (!permissionProvider.hasEditPermission(existingDatasource)) {
-                                        log.error(
-                                                "Trying to update datasource {} without edit permission",
-                                                existingDatasource.getName());
-                                        return Mono.error(new AppsmithException(
-                                                AppsmithError.ACL_NO_RESOURCE_FOUND,
-                                                FieldName.DATASOURCE,
-                                                existingDatasource.getId()));
-                                    }
-                                    datasourceStorage.setId(null);
-                                    // Don't update datasource config as the saved datasource is already configured by
-                                    // user
-                                    // for this instance
-                                    datasourceStorage.setDatasourceConfiguration(null);
-                                    datasourceStorage.setPluginId(null);
-                                    datasourceStorage.setEnvironmentId(environmentId);
-                                    Datasource newDatasource =
-                                            datasourceService.createDatasourceFromDatasourceStorage(datasourceStorage);
-                                    newDatasource.setPolicies(null);
-
-                                    copyNestedNonNullProperties(newDatasource, existingDatasource);
-                                    // Don't update the datasource configuration for already available datasources
-                                    existingDatasource.setDatasourceConfiguration(null);
-                                    return datasourceService.save(existingDatasource);
-                                } else {
-                                    // This is explicitly copied over from the map we created before
-                                    datasourceStorage.setPluginId(pluginMap.get(datasourceStorage.getPluginId()));
-                                    datasourceStorage.setWorkspaceId(workspace.getId());
-                                    datasourceStorage.setEnvironmentId(environmentId);
-
-                                    // Check if any decrypted fields are present for datasource
-                                    if (importedDoc.getDecryptedFields() != null
-                                            && importedDoc.getDecryptedFields().get(datasourceStorage.getName())
-                                                    != null) {
-
-                                        DecryptedSensitiveFields decryptedFields =
-                                                importedDoc.getDecryptedFields().get(datasourceStorage.getName());
-
-                                        updateAuthenticationDTO(datasourceStorage, decryptedFields);
-                                    }
-                                    return createUniqueDatasourceIfNotPresent(
-                                            existingDatasourcesFlux,
-                                            datasourceStorage,
-                                            workspace,
-                                            environmentId,
-                                            permissionProvider);
-                                }
-                            });
-                })
-                .collectMap(Datasource::getName, Datasource::getId)
-                .onErrorResume(error -> {
-                    log.error("Error importing datasources", error);
-                    return Mono.error(error);
-                })
-                .elapsed()
-                .map(tuple -> {
-                    log.debug("Time taken to import datasources: {} ms", tuple.getT1());
-                    return tuple.getT2();
-                });
-    }
-
     private Mono<Application> getImportApplicationMono(
             Application importedApplication,
-            String applicationId,
-            boolean appendToApp,
-            String workspaceId,
-            ImportApplicationPermissionProvider permissionProvider,
-            Mono<User> currUserMono,
-            Mono<List<CustomJSLibApplicationDTO>> installedJsLibsMono) {
+            ImportingMetaDTO importingMetaDTO,
+            MappedImportableResourcesDTO mappedImportableResourcesDTO,
+            Mono<User> currUserMono) {
         Mono<Application> importApplicationMono = Mono.just(importedApplication)
                 .map(application -> {
                     if (application.getApplicationVersion() == null) {
@@ -528,18 +363,16 @@ public class ImportApplicationServiceCEImpl implements ImportApplicationServiceC
                     application.setViewMode(false);
                     application.setForkWithConfiguration(null);
                     application.setExportWithConfiguration(null);
-                    application.setWorkspaceId(workspaceId);
+                    application.setWorkspaceId(importingMetaDTO.getWorkspaceId());
                     application.setIsPublic(null);
                     application.setPolicies(null);
                     application.setPages(null);
                     application.setPublishedPages(null);
                     return application;
                 })
-                .zipWith(installedJsLibsMono)
-                .map(tuple -> {
-                    Application application = tuple.getT1();
-                    List<CustomJSLibApplicationDTO> customJSLibApplicationDTOList = tuple.getT2();
-                    application.setUnpublishedCustomJSLibs(new HashSet<>(customJSLibApplicationDTOList));
+                .map(application -> {
+                    application.setUnpublishedCustomJSLibs(
+                            new HashSet<>(mappedImportableResourcesDTO.getInstalledJsLibsList()));
                     return application;
                 });
 
@@ -549,25 +382,29 @@ public class ImportApplicationServiceCEImpl implements ImportApplicationServiceC
             return application;
         });
 
-        if (StringUtils.isEmpty(applicationId)) {
+        if (StringUtils.isEmpty(importingMetaDTO.getApplicationId())) {
             importApplicationMono = importApplicationMono.flatMap(application -> {
                 return applicationPageService.createOrUpdateSuffixedApplication(application, application.getName(), 0);
             });
         } else {
             Mono<Application> existingApplicationMono = applicationService
-                    .findById(applicationId, permissionProvider.getRequiredPermissionOnTargetApplication())
+                    .findById(
+                            importingMetaDTO.getApplicationId(),
+                            importingMetaDTO.getPermissionProvider().getRequiredPermissionOnTargetApplication())
                     .switchIfEmpty(Mono.defer(() -> {
                         log.error(
                                 "No application found with id: {} and permission: {}",
-                                applicationId,
-                                permissionProvider.getRequiredPermissionOnTargetApplication());
+                                importingMetaDTO.getApplicationId(),
+                                importingMetaDTO.getPermissionProvider().getRequiredPermissionOnTargetApplication());
                         return Mono.error(new AppsmithException(
-                                AppsmithError.ACL_NO_RESOURCE_FOUND, FieldName.APPLICATION, applicationId));
+                                AppsmithError.ACL_NO_RESOURCE_FOUND,
+                                FieldName.APPLICATION,
+                                importingMetaDTO.getApplicationId()));
                     }))
                     .cache();
 
             // this can be a git sync, import page from template, update app with json, restore snapshot
-            if (appendToApp) { // we don't need to do anything with the imported application
+            if (importingMetaDTO.getAppendToApp()) { // we don't need to do anything with the imported application
                 importApplicationMono = existingApplicationMono;
             } else {
                 importApplicationMono = importApplicationMono
@@ -623,628 +460,6 @@ public class ImportApplicationServiceCEImpl implements ImportApplicationServiceC
                 });
     }
 
-    private Mono<Map<String, String>> getPluginMapMono() {
-        return pluginRepository
-                .findAll()
-                .collectList()
-                .map(plugins -> {
-                    Map<String, String> pluginMap = new HashMap<>();
-                    plugins.forEach(plugin -> {
-                        final String pluginReference = StringUtils.isEmpty(plugin.getPluginName())
-                                ? plugin.getPackageName()
-                                : plugin.getPluginName();
-                        pluginMap.put(pluginReference, plugin.getId());
-                    });
-                    return pluginMap;
-                })
-                .elapsed()
-                .map(tuples -> {
-                    log.debug("time to get plugin map: {}", tuples.getT1());
-                    return tuples.getT2();
-                });
-    }
-
-    private Mono<Map<String, NewPage>> getPageNameMapMono(
-            Mono<Tuple2<List<NewPage>, Map<String, String>>> importedNewPagesMono) {
-        return importedNewPagesMono.map(objects -> {
-            Map<String, NewPage> pageNameMap = new HashMap<>();
-            objects.getT1().forEach(newPage -> {
-                // Save the map of pageName and NewPage
-                if (newPage.getUnpublishedPage() != null
-                        && newPage.getUnpublishedPage().getName() != null) {
-                    pageNameMap.put(newPage.getUnpublishedPage().getName(), newPage);
-                }
-                if (newPage.getPublishedPage() != null
-                        && newPage.getPublishedPage().getName() != null) {
-                    pageNameMap.put(newPage.getPublishedPage().getName(), newPage);
-                }
-            });
-            return pageNameMap;
-        });
-    }
-
-    private Mono<Tuple2<ImportedActionAndCollectionMapsDTO, ImportActionResultDTO>> setCollectionIdToActionsMono(
-            Mono<ImportActionResultDTO> importActionsMono,
-            Mono<ImportActionCollectionResultDTO> importActionCollectionMono,
-            Mono<Application> importApplicationMono,
-            String applicationId,
-            boolean appendToApp) {
-        return Mono.zip(importActionsMono, importActionCollectionMono, importApplicationMono)
-                .flatMap(objects -> {
-                    ImportActionResultDTO importActionResultDTO = objects.getT1();
-                    ImportActionCollectionResultDTO importActionCollectionResultDTO = objects.getT2();
-                    Application importedApplication1 = objects.getT3();
-
-                    List<String> savedCollectionIds = importActionCollectionResultDTO.getSavedActionCollectionIds();
-                    log.info(
-                            "Action collections imported. applicationId {}, result: {}",
-                            importedApplication1.getId(),
-                            importActionCollectionResultDTO.getGist());
-                    return newActionService
-                            .updateActionsWithImportedCollectionIds(
-                                    importActionCollectionResultDTO, importActionResultDTO)
-                            .flatMap(actionAndCollectionMapsDTO -> {
-                                log.info(
-                                        "Updated actions with imported collection ids. applicationId {}",
-                                        importedApplication1.getId());
-                                // Updating the existing application for git-sync
-                                // During partial import/appending to the existing application keep the resources
-                                // attached to the application:
-                                // Delete the invalid resources (which are not the part of applicationJsonDTO) in
-                                // the git flow only
-                                if (!StringUtils.isEmpty(applicationId) && !appendToApp) {
-                                    // Remove unwanted action collections
-                                    Set<String> invalidCollectionIds = new HashSet<>();
-                                    for (ActionCollection collection :
-                                            importActionCollectionResultDTO.getExistingActionCollections()) {
-                                        if (!savedCollectionIds.contains(collection.getId())) {
-                                            invalidCollectionIds.add(collection.getId());
-                                        }
-                                    }
-                                    log.info(
-                                            "Deleting {} action collections which are no more used",
-                                            invalidCollectionIds.size());
-                                    return Flux.fromIterable(invalidCollectionIds)
-                                            .flatMap(collectionId -> actionCollectionService
-                                                    .deleteWithoutPermissionUnpublishedActionCollection(collectionId)
-                                                    // return an empty collection so that the filter can remove it from
-                                                    // the list
-                                                    .onErrorResume(throwable -> {
-                                                        log.debug(
-                                                                "Failed to delete collection with id {} during import",
-                                                                collectionId);
-                                                        log.error(throwable.getMessage());
-                                                        return Mono.empty();
-                                                    }))
-                                            .then()
-                                            .thenReturn(actionAndCollectionMapsDTO);
-                                }
-                                return Mono.just(actionAndCollectionMapsDTO);
-                            })
-                            .zipWith(Mono.just(importActionResultDTO));
-                })
-                .onErrorResume(error -> {
-                    log.error("Error while updating collection id to actions", error);
-                    return Mono.error(error);
-                });
-    }
-
-    private Mono<List<NewPage>> setActionIdInPages(
-            Mono<Tuple2<ImportedActionAndCollectionMapsDTO, ImportActionResultDTO>> updateCollectionIdToActionsMono,
-            Mono<Tuple2<List<NewPage>, Map<String, String>>> importedNewPagesMono,
-            String branchName) {
-        return Mono.zip(updateCollectionIdToActionsMono, importedNewPagesMono)
-                .flatMap(objects -> {
-                    ImportedActionAndCollectionMapsDTO actionAndCollectionMapsDTO =
-                            objects.getT1().getT1();
-                    ImportActionResultDTO importActionResultDTO =
-                            objects.getT1().getT2();
-                    List<NewPage> newPages = objects.getT2().getT1();
-                    return Flux.fromIterable(newPages)
-                            .flatMap(newPage -> {
-                                if (newPage.getDefaultResources() != null) {
-                                    newPage.getDefaultResources().setBranchName(branchName);
-                                }
-                                return mapActionAndCollectionIdWithPageLayout(
-                                        newPage,
-                                        importActionResultDTO.getActionIdMap(),
-                                        actionAndCollectionMapsDTO.getUnpublishedActionIdToCollectionIdMap(),
-                                        actionAndCollectionMapsDTO.getPublishedActionIdToCollectionIdMap());
-                            })
-                            .collectList()
-                            .flatMapMany(newPageService::saveAll)
-                            .collectList();
-                })
-                .onErrorResume(throwable -> {
-                    log.error("Failed to set action ids in pages", throwable);
-                    return Mono.error(throwable);
-                });
-    }
-
-    private Mono<ImportActionCollectionResultDTO> createImportActionCollectionMono(
-            List<ActionCollection> importedActionCollectionList,
-            Mono<Application> importApplicationMono,
-            Mono<Tuple2<List<NewPage>, Map<String, String>>> importedNewPagesMono,
-            Mono<ImportActionResultDTO> importActionsMono,
-            Mono<Map<String, String>> pluginMapMno,
-            Mono<Map<String, NewPage>> pageNameMapMono,
-            ImportApplicationPermissionProvider permissionProvider,
-            boolean appendToApp,
-            String branchName) {
-        Mono<List<ActionCollection>> importedActionCollectionMono = Mono.just(importedActionCollectionList);
-
-        if (appendToApp) {
-            importedActionCollectionMono = importedActionCollectionMono
-                    .zipWith(importedNewPagesMono)
-                    .map(objects -> {
-                        List<ActionCollection> importedActionCollectionList1 = objects.getT1();
-                        List<NewPage> importedNewPages = objects.getT2().getT1();
-                        Map<String, String> newToOldNameMap = objects.getT2().getT2();
-
-                        for (NewPage newPage : importedNewPages) {
-                            String newPageName = newPage.getUnpublishedPage().getName();
-                            String oldPageName = newToOldNameMap.get(newPageName);
-
-                            if (!newPageName.equals(oldPageName)) {
-                                renamePageInActionCollections(importedActionCollectionList1, oldPageName, newPageName);
-                            }
-                        }
-                        return importedActionCollectionList1;
-                    });
-        }
-
-        return Mono.zip(
-                        importActionsMono,
-                        importApplicationMono,
-                        importedActionCollectionMono,
-                        pluginMapMno,
-                        pageNameMapMono)
-                .flatMap(objects -> {
-                    log.info("Importing action collections");
-                    return actionCollectionService.importActionCollections(
-                            objects.getT1(),
-                            objects.getT2(),
-                            branchName,
-                            objects.getT3(),
-                            objects.getT4(),
-                            objects.getT5(),
-                            permissionProvider);
-                })
-                .onErrorResume(throwable -> {
-                    log.error("Error importing action collections", throwable);
-                    return Mono.error(throwable);
-                });
-    }
-
-    Mono<ImportActionResultDTO> getImportActionsMono(
-            ApplicationJson importedDoc,
-            Mono<Tuple2<List<NewPage>, Map<String, String>>> importedNewPagesMono,
-            Mono<Application> importApplicationMono,
-            Mono<Map<String, NewPage>> pageNameMapMono,
-            Mono<Map<String, String>> pluginMapMno,
-            boolean appendToApp,
-            Workspace workspace,
-            String applicationId,
-            String branchName,
-            ImportApplicationPermissionProvider permissionProvider) {
-        List<NewAction> importedNewActionList = importedDoc.getActionList();
-
-        final Flux<Datasource> existingDatasourceFlux = datasourceService
-                .getAllByWorkspaceIdWithStorages(workspace.getId(), Optional.empty())
-                .cache();
-
-        Mono<List<Datasource>> existingDatasourceMono =
-                getExistingDatasourceMono(applicationId, existingDatasourceFlux);
-        Mono<Map<String, String>> datasourceMapMono = importDatasources(
-                importedDoc,
-                existingDatasourceMono,
-                existingDatasourceFlux,
-                workspace,
-                pluginMapMno,
-                permissionProvider);
-
-        Mono<List<NewAction>> importedNewActionMono = Mono.just(importedNewActionList);
-        if (appendToApp) {
-            importedNewActionMono = importedNewActionMono
-                    .zipWith(importedNewPagesMono)
-                    .map(objects -> {
-                        List<NewAction> importedNewActionList1 = objects.getT1();
-                        List<NewPage> importedNewPages = objects.getT2().getT1();
-                        Map<String, String> newToOldNameMap = objects.getT2().getT2();
-
-                        for (NewPage newPage : importedNewPages) {
-                            String newPageName = newPage.getUnpublishedPage().getName();
-                            String oldPageName = newToOldNameMap.get(newPageName);
-
-                            if (!newPageName.equals(oldPageName)) {
-                                renamePageInActions(importedNewActionList1, oldPageName, newPageName);
-                            }
-                        }
-                        return importedNewActionList1;
-                    });
-        }
-
-        return Mono.zip(importedNewActionMono, importApplicationMono, pageNameMapMono, pluginMapMno, datasourceMapMono)
-                .flatMap(objects -> newActionService.importActions(
-                        objects.getT1(),
-                        objects.getT2(),
-                        branchName,
-                        objects.getT3(),
-                        objects.getT4(),
-                        objects.getT5(),
-                        permissionProvider))
-                .flatMap(importActionResultDTO -> {
-                    log.info("Actions imported. result: {}", importActionResultDTO.getGist());
-                    // Updating the existing application for git-sync
-                    // During partial import/appending to the existing application keep the resources
-                    // attached to the application:
-                    // Delete the invalid resources (which are not the part of applicationJsonDTO) in
-                    // the git flow only
-                    if (!StringUtils.isEmpty(applicationId)
-                            && !appendToApp
-                            && CollectionUtils.isNotEmpty(importActionResultDTO.getExistingActions())) {
-                        // Remove unwanted actions
-                        Set<String> invalidActionIds = new HashSet<>();
-                        for (NewAction action : importActionResultDTO.getExistingActions()) {
-                            if (!importActionResultDTO.getImportedActionIds().contains(action.getId())) {
-                                invalidActionIds.add(action.getId());
-                            }
-                        }
-                        log.info("Deleting {} actions which are no more used", invalidActionIds.size());
-                        return Flux.fromIterable(invalidActionIds)
-                                .flatMap(actionId -> newActionService
-                                        .deleteUnpublishedAction(actionId)
-                                        // return an empty action so that the filter can remove it from the list
-                                        .onErrorResume(throwable -> {
-                                            log.debug("Failed to delete action with id {} during import", actionId);
-                                            log.error(throwable.getMessage());
-                                            return Mono.empty();
-                                        }))
-                                .then()
-                                .thenReturn(importActionResultDTO);
-                    }
-                    return Mono.just(importActionResultDTO);
-                })
-                .onErrorResume(throwable -> {
-                    log.error("Error while importing actions and deleting unused ones", throwable);
-                    return Mono.error(throwable);
-                });
-    }
-
-    private Mono<List<ApplicationPage>> importUnpublishedPages(
-            List<ApplicationPage> editModeApplicationPages,
-            boolean appendToApp,
-            Mono<Application> importApplicationMono,
-            Mono<Tuple2<List<NewPage>, Map<String, String>>> importedNewPagesMono) {
-        Mono<List<ApplicationPage>> unpublishedPagesMono = Mono.just(editModeApplicationPages);
-        if (appendToApp) {
-            unpublishedPagesMono = unpublishedPagesMono
-                    .zipWith(importApplicationMono)
-                    .map(objects -> {
-                        Application application = objects.getT2();
-                        List<ApplicationPage> applicationPages = objects.getT1();
-                        applicationPages.addAll(application.getPages());
-                        return applicationPages;
-                    })
-                    .zipWith(importedNewPagesMono)
-                    .map(objects -> {
-                        List<ApplicationPage> unpublishedPages = objects.getT1();
-                        Map<String, String> newToOldNameMap = objects.getT2().getT2();
-                        List<NewPage> importedPages = objects.getT2().getT1();
-                        for (NewPage newPage : importedPages) {
-                            // we need to map the newly created page with old name
-                            // because other related resources e.g. actions will refer the page with old name
-                            String newPageName = newPage.getUnpublishedPage().getName();
-                            if (newToOldNameMap.containsKey(newPageName)) {
-                                String oldPageName = newToOldNameMap.get(newPageName);
-                                unpublishedPages.stream()
-                                        .filter(applicationPage -> oldPageName.equals(applicationPage.getId()))
-                                        .findAny()
-                                        .ifPresent(applicationPage -> applicationPage.setId(newPageName));
-                            }
-                        }
-                        return unpublishedPages;
-                    });
-        }
-        return unpublishedPagesMono;
-    }
-
-    private Mono<Tuple2<List<NewPage>, Map<String, String>>> getImportNewPagesMono(
-            List<NewPage> importedNewPageList,
-            Mono<List<NewPage>> existingPagesMono,
-            Mono<Application> importApplicationMono,
-            boolean appendToApp,
-            String branchName,
-            ImportApplicationPermissionProvider permissionProvider) {
-        return Mono.just(importedNewPageList)
-                .zipWith(existingPagesMono)
-                .map(objects -> {
-                    List<NewPage> importedNewPages = objects.getT1();
-                    List<NewPage> existingPages = objects.getT2();
-                    Map<String, String> newToOldNameMap;
-                    if (appendToApp) {
-                        newToOldNameMap = updateNewPagesBeforeMerge(existingPages, importedNewPages);
-                    } else {
-                        newToOldNameMap = Map.of();
-                    }
-                    return Tuples.of(importedNewPages, newToOldNameMap);
-                })
-                .zipWith(importApplicationMono)
-                .flatMap(objects -> {
-                    List<NewPage> importedNewPages = objects.getT1().getT1();
-                    Map<String, String> newToOldNameMap = objects.getT1().getT2();
-                    Application application = objects.getT2();
-                    return importAndSavePages(
-                                    importedNewPages, application, branchName, existingPagesMono, permissionProvider)
-                            .collectList()
-                            .zipWith(Mono.just(newToOldNameMap));
-                })
-                .onErrorResume(throwable -> {
-                    log.error("Error importing pages", throwable);
-                    return Mono.error(throwable);
-                })
-                .elapsed()
-                .map(objects -> {
-                    log.debug("time to import {} pages: {}", objects.getT2().size(), objects.getT1());
-                    return objects.getT2();
-                });
-    }
-
-    Mono<Application> savePagesToApplicationMono(
-            Application importedApplication,
-            Mono<Map<String, NewPage>> pageNameMapMono,
-            Mono<Application> applicationMono,
-            boolean appendToApp,
-            String applicationId,
-            Mono<List<NewPage>> existingPagesMono,
-            Mono<Tuple2<List<NewPage>, Map<String, String>>> importedNewPagesMono) {
-
-        List<ApplicationPage> editModeApplicationPages = importedApplication.getPages();
-        List<ApplicationPage> publishedModeApplicationPages = importedApplication.getPublishedPages();
-
-        Mono<List<ApplicationPage>> unpublishedPagesMono =
-                importUnpublishedPages(editModeApplicationPages, appendToApp, applicationMono, importedNewPagesMono);
-
-        Mono<List<ApplicationPage>> publishedPagesMono = Mono.just(publishedModeApplicationPages);
-
-        Mono<Map<ResourceModes, List<ApplicationPage>>> applicationPagesMono = Mono.zip(
-                        unpublishedPagesMono, publishedPagesMono, pageNameMapMono, applicationMono)
-                .map(objects -> {
-                    List<ApplicationPage> unpublishedPages = objects.getT1();
-                    List<ApplicationPage> publishedPages = objects.getT2();
-                    Map<String, NewPage> pageNameMap = objects.getT3();
-                    Application savedApp = objects.getT4();
-
-                    log.debug("New pages imported for application: {}", savedApp.getId());
-                    Map<ResourceModes, List<ApplicationPage>> applicationPages = new HashMap<>();
-                    applicationPages.put(EDIT, unpublishedPages);
-                    applicationPages.put(VIEW, publishedPages);
-
-                    Iterator<ApplicationPage> unpublishedPageItr = unpublishedPages.iterator();
-                    while (unpublishedPageItr.hasNext()) {
-                        ApplicationPage applicationPage = unpublishedPageItr.next();
-                        NewPage newPage = pageNameMap.get(applicationPage.getId());
-                        if (newPage == null) {
-                            if (appendToApp) {
-                                // Don't remove the page reference if doing the partial import and appending
-                                // to the existing application
-                                continue;
-                            }
-                            log.debug(
-                                    "Unable to find the page during import for appId {}, with name {}",
-                                    applicationId,
-                                    applicationPage.getId());
-                            unpublishedPageItr.remove();
-                        } else {
-                            applicationPage.setId(newPage.getId());
-                            applicationPage.setDefaultPageId(
-                                    newPage.getDefaultResources().getPageId());
-                            // Keep the existing page as the default one
-                            if (appendToApp) {
-                                applicationPage.setIsDefault(false);
-                            }
-                        }
-                    }
-
-                    Iterator<ApplicationPage> publishedPagesItr;
-                    // Remove the newly added pages from merge app flow. Keep only the existing page from the old app
-                    if (appendToApp) {
-                        List<String> existingPagesId = savedApp.getPublishedPages().stream()
-                                .map(applicationPage -> applicationPage.getId())
-                                .collect(Collectors.toList());
-                        List<ApplicationPage> publishedApplicationPages = publishedPages.stream()
-                                .filter(applicationPage -> existingPagesId.contains(applicationPage.getId()))
-                                .collect(Collectors.toList());
-                        applicationPages.replace(VIEW, publishedApplicationPages);
-                        publishedPagesItr = publishedApplicationPages.iterator();
-                    } else {
-                        publishedPagesItr = publishedPages.iterator();
-                    }
-                    while (publishedPagesItr.hasNext()) {
-                        ApplicationPage applicationPage = publishedPagesItr.next();
-                        NewPage newPage = pageNameMap.get(applicationPage.getId());
-                        if (newPage == null) {
-                            log.debug(
-                                    "Unable to find the page during import for appId {}, with name {}",
-                                    applicationId,
-                                    applicationPage.getId());
-                            if (!appendToApp) {
-                                publishedPagesItr.remove();
-                            }
-                        } else {
-                            applicationPage.setId(newPage.getId());
-                            applicationPage.setDefaultPageId(
-                                    newPage.getDefaultResources().getPageId());
-                            if (appendToApp) {
-                                applicationPage.setIsDefault(false);
-                            }
-                        }
-                    }
-
-                    return applicationPages;
-                });
-
-        if (!StringUtils.isEmpty(applicationId) && !appendToApp) {
-            applicationPagesMono = applicationPagesMono
-                    .zipWith(existingPagesMono)
-                    .flatMap(objects -> {
-                        Map<ResourceModes, List<ApplicationPage>> applicationPages = objects.getT1();
-                        List<NewPage> existingPagesList = objects.getT2();
-                        Set<String> validPageIds = applicationPages.get(EDIT).stream()
-                                .map(ApplicationPage::getId)
-                                .collect(Collectors.toSet());
-
-                        validPageIds.addAll(applicationPages.get(VIEW).stream()
-                                .map(ApplicationPage::getId)
-                                .collect(Collectors.toSet()));
-
-                        Set<String> invalidPageIds = new HashSet<>();
-                        for (NewPage newPage : existingPagesList) {
-                            if (!validPageIds.contains(newPage.getId())) {
-                                invalidPageIds.add(newPage.getId());
-                            }
-                        }
-
-                        // Delete the pages which were removed during git merge operation
-                        // This does not apply to the traditional import via file approach
-                        return Flux.fromIterable(invalidPageIds)
-                                .flatMap(applicationPageService::deleteWithoutPermissionUnpublishedPage)
-                                .flatMap(page -> newPageService
-                                        .archiveWithoutPermissionById(page.getId())
-                                        .onErrorResume(e -> {
-                                            log.debug(
-                                                    "Unable to archive page {} with error {}",
-                                                    page.getId(),
-                                                    e.getMessage());
-                                            return Mono.empty();
-                                        }))
-                                .then()
-                                .thenReturn(applicationPages);
-                    });
-        }
-        return applicationMono.zipWith(applicationPagesMono).map(objects -> {
-            Application application = objects.getT1();
-            Map<ResourceModes, List<ApplicationPage>> applicationPages = objects.getT2();
-            application.setPages(applicationPages.get(EDIT));
-            application.setPublishedPages(applicationPages.get(VIEW));
-            return application;
-        });
-    }
-
-    private Mono<List<CustomJSLibApplicationDTO>> getCustomJslibImportMono(List<CustomJSLib> customJSLibs) {
-        if (customJSLibs == null) {
-            customJSLibs = new ArrayList<>();
-        }
-
-        ensureXmlParserPresenceInCustomJsLibList(customJSLibs);
-
-        return Flux.fromIterable(customJSLibs)
-                .flatMap(customJSLib -> {
-                    customJSLib.setId(null);
-                    customJSLib.setCreatedAt(null);
-                    customJSLib.setUpdatedAt(null);
-                    return customJSLibService.persistCustomJSLibMetaDataIfDoesNotExistAndGetDTO(customJSLib, false);
-                })
-                .collectList()
-                .elapsed()
-                .map(objects -> {
-                    log.debug("time to import custom jslibs: {}", objects.getT1());
-                    return objects.getT2();
-                })
-                .onErrorResume(e -> {
-                    log.error("Error importing custom jslibs", e);
-                    return Mono.error(e);
-                });
-    }
-
-    /**
-     * This method takes customJSLibList from application JSON, checks if an entry for XML parser exists,
-     * otherwise adds the entry.
-     * This has been done to add the xmlParser entry in imported application as appsmith is stopping native support
-     * for xml parser.
-     * Read More: https://github.com/appsmithorg/appsmith/pull/28012
-     *
-     * @param customJSLibList
-     */
-    public void ensureXmlParserPresenceInCustomJsLibList(List<CustomJSLib> customJSLibList) {
-        boolean isXmlParserLibFound = false;
-        for (CustomJSLib customJSLib : customJSLibList) {
-            if (!customJSLib.getUidString().equals(ApplicationConstants.XML_PARSER_LIBRARY_UID)) {
-                continue;
-            }
-
-            isXmlParserLibFound = true;
-            break;
-        }
-
-        if (!isXmlParserLibFound) {
-            CustomJSLib xmlParserJsLib = ApplicationConstants.getDefaultParserCustomJsLibCompatibilityDTO();
-            customJSLibList.add(xmlParserJsLib);
-        }
-    }
-
-    private Mono<List<Datasource>> getExistingDatasourceMono(String applicationId, Flux<Datasource> datasourceFlux) {
-        Mono<List<Datasource>> existingDatasourceMono;
-        // Check if the request is to hydrate the application to DB for particular branch
-        // Application id will be present for GIT sync
-        if (!StringUtils.isEmpty(applicationId)) {
-            // No need to hydrate the datasource as we expect user will configure the datasource
-            existingDatasourceMono = datasourceFlux.collectList();
-        } else {
-            existingDatasourceMono = Mono.just(new ArrayList<>());
-        }
-        return existingDatasourceMono;
-    }
-
-    private Mono<List<NewPage>> getImportActionsAndActionCollectionsMono(
-            ApplicationJson importedDoc,
-            Mono<Application> applicationMono,
-            Mono<Tuple2<List<NewPage>, Map<String, String>>> importedNewPagesMono,
-            Mono<Map<String, String>> pluginMapMno,
-            Mono<Map<String, NewPage>> pageNameMapMono,
-            boolean appendToApp,
-            String branchName,
-            String applicationId,
-            Workspace workspace,
-            ImportApplicationPermissionProvider permissionProvider) {
-        Mono<ImportActionResultDTO> importActionsMono = getImportActionsMono(
-                        importedDoc,
-                        importedNewPagesMono,
-                        applicationMono,
-                        pageNameMapMono,
-                        pluginMapMno,
-                        appendToApp,
-                        workspace,
-                        applicationId,
-                        branchName,
-                        permissionProvider)
-                .cache();
-
-        List<ActionCollection> importedActionCollectionList =
-                CollectionUtils.isEmpty(importedDoc.getActionCollectionList())
-                        ? new ArrayList<>()
-                        : importedDoc.getActionCollectionList();
-
-        Mono<ImportActionCollectionResultDTO> importActionCollectionMono = createImportActionCollectionMono(
-                importedActionCollectionList,
-                applicationMono,
-                importedNewPagesMono,
-                importActionsMono,
-                pluginMapMno,
-                pageNameMapMono,
-                permissionProvider,
-                appendToApp,
-                branchName);
-
-        Mono<Tuple2<ImportedActionAndCollectionMapsDTO, ImportActionResultDTO>> updateCollectionIdToActionsMono =
-                setCollectionIdToActionsMono(
-                        importActionsMono, importActionCollectionMono, applicationMono, applicationId, appendToApp);
-
-        Mono<List<NewPage>> updateActionsInPages =
-                setActionIdInPages(updateCollectionIdToActionsMono, importedNewPagesMono, branchName);
-        return updateActionsInPages;
-    }
-
     /**
      * This function will take the application reference object to hydrate the application in mongoDB
      *
@@ -1281,8 +496,12 @@ public class ImportApplicationServiceCEImpl implements ImportApplicationServiceC
                     AppsmithError.VALIDATION_FAILURE, "Field '" + errorField + "' is missing in the JSON."));
         }
 
+        ImportingMetaDTO importingMetaDTO =
+                new ImportingMetaDTO(workspaceId, applicationId, branchName, appendToApp, permissionProvider);
+
+        MappedImportableResourcesDTO mappedImportableResourcesDTO = new MappedImportableResourcesDTO();
+
         Application importedApplication = importedDoc.getExportedApplication();
-        List<NewPage> importedNewPageList = importedDoc.getPageList();
 
         Mono<Workspace> workspaceMono = workspaceService
                 .findById(workspaceId, permissionProvider.getRequiredPermissionOnTargetWorkspace())
@@ -1293,93 +512,90 @@ public class ImportApplicationServiceCEImpl implements ImportApplicationServiceC
                             permissionProvider.getRequiredPermissionOnTargetWorkspace());
                     return Mono.error(new AppsmithException(
                             AppsmithError.ACL_NO_RESOURCE_FOUND, FieldName.WORKSPACE, workspaceId));
-                }));
+                }))
+                .cache();
 
         Mono<User> currUserMono = sessionUserService.getCurrentUser().cache();
-        Mono<List<CustomJSLibApplicationDTO>> installedJsLibsMono =
-                getCustomJslibImportMono(importedDoc.getCustomJSLibList());
-        Mono<Map<String, String>> pluginMapMno = getPluginMapMono().cache();
+        Mono<List<CustomJSLib>> installedJsLibsMono = customJSLibImportableService.importEntities(
+                importingMetaDTO, mappedImportableResourcesDTO, null, null, applicationJson);
+        Mono<List<Plugin>> installedPluginsMono = pluginImportableService.importEntities(
+                importingMetaDTO, mappedImportableResourcesDTO, workspaceMono, null, applicationJson);
 
         // Start the stopwatch to log the execution time
         Stopwatch stopwatch = new Stopwatch(AnalyticsEvents.IMPORT.getEventName());
-        Mono<Application> importedApplicationMono = workspaceMono
-                .flatMap(workspace -> {
-                    Mono<Application> applicationMono = getImportApplicationMono(
-                                    importedApplication,
-                                    applicationId,
-                                    appendToApp,
-                                    workspaceId,
-                                    permissionProvider,
-                                    currUserMono,
-                                    installedJsLibsMono)
-                            .flatMap(application -> importThemes(application, importedDoc, appendToApp))
-                            .cache();
+        final Mono<Application> importedApplicationMono = installedJsLibsMono
+                .then(getImportApplicationMono(
+                        importedApplication, importingMetaDTO, mappedImportableResourcesDTO, currUserMono))
+                .cache();
 
-                    // Import and save pages, also update the pages related fields in saved application
-                    assert importedNewPageList != null : "Unable to find pages in the imported application";
+        Mono<List<Theme>> importedThemesMono = themeImportableService.importEntities(
+                importingMetaDTO,
+                mappedImportableResourcesDTO,
+                workspaceMono,
+                importedApplicationMono,
+                applicationJson);
 
-                    // For git-sync this will not be empty
-                    Mono<List<NewPage>> existingPagesMono = applicationMono
-                            .flatMap(application -> newPageService
-                                    .findNewPagesByApplicationId(application.getId(), Optional.empty())
-                                    .collectList())
-                            .cache();
+        Mono<List<NewPage>> importedPagesMono = newPageImportableService.importEntities(
+                importingMetaDTO,
+                mappedImportableResourcesDTO,
+                workspaceMono,
+                importedApplicationMono,
+                applicationJson);
 
-                    Mono<Tuple2<List<NewPage>, Map<String, String>>> importedNewPagesMono = getImportNewPagesMono(
-                                    importedNewPageList,
-                                    existingPagesMono,
-                                    applicationMono,
-                                    appendToApp,
-                                    branchName,
-                                    permissionProvider)
-                            .cache();
+        Mono<List<Datasource>> importedDatasourcesMono = datasourceImportableService.importEntities(
+                importingMetaDTO,
+                mappedImportableResourcesDTO,
+                workspaceMono,
+                importedApplicationMono,
+                applicationJson);
 
-                    Mono<Map<String, NewPage>> pageNameMapMono =
-                            getPageNameMapMono(importedNewPagesMono).cache();
+        Mono<List<NewAction>> importedNewActionsMono = newActionImportableService.importEntities(
+                importingMetaDTO,
+                mappedImportableResourcesDTO,
+                workspaceMono,
+                importedApplicationMono,
+                applicationJson);
 
-                    applicationMono = savePagesToApplicationMono(
-                                    importedApplication,
-                                    pageNameMapMono,
-                                    applicationMono,
-                                    appendToApp,
-                                    applicationId,
-                                    existingPagesMono,
-                                    importedNewPagesMono)
-                            .cache();
+        Mono<List<ActionCollection>> importedActionCollectionsMono = actionCollectionImportableService.importEntities(
+                importingMetaDTO,
+                mappedImportableResourcesDTO,
+                workspaceMono,
+                importedApplicationMono,
+                applicationJson);
 
-                    // this will trigger all the dependent monos
-                    Mono<List<NewPage>> importActionAndActionCollections = getImportActionsAndActionCollectionsMono(
-                            importedDoc,
-                            applicationMono,
-                            importedNewPagesMono,
-                            pluginMapMno,
-                            pageNameMapMono,
-                            appendToApp,
-                            branchName,
-                            applicationId,
-                            workspace,
-                            permissionProvider);
+        Mono<Application> importMono = importedApplicationMono
+                .then(installedPluginsMono)
+                .then(importedThemesMono)
+                .then(importedPagesMono)
+                .then(importedDatasourcesMono)
+                .then(importedNewActionsMono)
+                .then(importedActionCollectionsMono)
+                .then(importedApplicationMono)
+                .flatMap(application -> {
+                    return newActionImportableService
+                            .updateImportedEntities(application, importingMetaDTO, mappedImportableResourcesDTO)
+                            .then(newPageImportableService.updateImportedEntities(
+                                    application, importingMetaDTO, mappedImportableResourcesDTO))
+                            .thenReturn(application);
+                })
+                .flatMap(application -> {
+                    log.info("Imported application with id {}", application.getId());
+                    // Need to update the application object with updated pages and publishedPages
+                    Application updateApplication = new Application();
+                    updateApplication.setPages(application.getPages());
+                    updateApplication.setPublishedPages(application.getPublishedPages());
 
-                    return importActionAndActionCollections
-                            .then(applicationMono)
-                            .flatMap(application -> {
-                                log.info("Imported application with id {}", application.getId());
-                                // Need to update the application object with updated pages and publishedPages
-                                Application updateApplication = new Application();
-                                updateApplication.setPages(application.getPages());
-                                updateApplication.setPublishedPages(application.getPublishedPages());
-                                return applicationService.update(application.getId(), updateApplication);
-                            })
-                            .onErrorResume(throwable -> {
-                                String errorMessage = ImportExportUtils.getErrorMessage(throwable);
-                                log.error("Error importing application. Error: {}", errorMessage, throwable);
-                                return Mono.error(new AppsmithException(
-                                        AppsmithError.GENERIC_JSON_IMPORT_ERROR, workspaceId, errorMessage));
-                            });
+                    return applicationService.update(application.getId(), updateApplication);
+                })
+                .onErrorResume(throwable -> {
+                    String errorMessage = ImportExportUtils.getErrorMessage(throwable);
+                    log.error("Error importing application. Error: {}", errorMessage, throwable);
+                    return Mono.error(
+                            new AppsmithException(AppsmithError.GENERIC_JSON_IMPORT_ERROR, workspaceId, errorMessage));
                 })
                 .as(transactionalOperator::transactional);
 
-        final Mono<Application> resultMono = importedApplicationMono
+        final Mono<Application> resultMono = importMono
                 .flatMap(application ->
                         sendImportExportApplicationAnalyticsEvent(application.getId(), AnalyticsEvents.IMPORT))
                 .zipWith(currUserMono)
@@ -1419,463 +635,12 @@ public class ImportApplicationServiceCEImpl implements ImportApplicationServiceC
 
         // Import Application is currently a slow API because it needs to import and create application, pages, actions
         // and action collection. This process may take time and the client may cancel the request. This leads to the
-        // flow
-        // getting stopped midway producing corrupted objects in DB. The following ensures that even though the client
-        // may have
-        // cancelled the flow, the importing the application should proceed uninterrupted and whenever the user
-        // refreshes
-        // the page, the imported application is available and is in sane state.
+        // flow getting stopped midway producing corrupted objects in DB. The following ensures that even though the
+        // client may have refreshes the page, the imported application is available and is in sane state.
         // To achieve this, we use a synchronous sink which does not take subscription cancellations into account. This
         // means that even if the subscriber has cancelled its subscription, the create method still generates its
         // event.
         return Mono.create(sink -> resultMono.subscribe(sink::success, sink::error, null, sink.currentContext()));
-    }
-
-    private void renamePageInActions(List<NewAction> newActionList, String oldPageName, String newPageName) {
-        for (NewAction newAction : newActionList) {
-            if (newAction.getUnpublishedAction().getPageId().equals(oldPageName)) {
-                newAction.getUnpublishedAction().setPageId(newPageName);
-            }
-        }
-    }
-
-    private void renamePageInActionCollections(
-            List<ActionCollection> actionCollectionList, String oldPageName, String newPageName) {
-        for (ActionCollection actionCollection : actionCollectionList) {
-            if (actionCollection.getUnpublishedCollection().getPageId().equals(oldPageName)) {
-                actionCollection.getUnpublishedCollection().setPageId(newPageName);
-            }
-        }
-    }
-
-    /**
-     * This function will respond with unique suffixed number for the entity to avoid duplicate names
-     *
-     * @param sourceEntity for which the suffixed number is required to avoid duplication
-     * @param workspaceId  workspace in which entity should be searched
-     * @return next possible number in case of duplication
-     */
-    private Mono<String> getUniqueSuffixForDuplicateNameEntity(BaseDomain sourceEntity, String workspaceId) {
-        if (sourceEntity != null) {
-            return sequenceService
-                    .getNextAsSuffix(sourceEntity.getClass(), " for workspace with _id : " + workspaceId)
-                    .map(sequenceNumber -> {
-                        // sequence number will be empty if no duplicate is found
-                        return sequenceNumber.isEmpty() ? " #1" : " #" + sequenceNumber.trim();
-                    });
-        }
-        return Mono.just("");
-    }
-
-    /**
-     * Method to
-     * - save imported pages
-     * - update the mongoEscapedWidgets if present in the page
-     * - set the policies for the page
-     * - update default resource ids along with branch-name if the application is connected to git
-     *
-     * @param pages         pagelist extracted from the imported JSON file
-     * @param application   saved application where pages needs to be added
-     * @param branchName    to which branch pages should be imported if application is connected to git
-     * @param existingPages existing pages in DB if the application is connected to git
-     * @return flux of saved pages in DB
-     */
-    private Flux<NewPage> importAndSavePages(
-            List<NewPage> pages,
-            Application application,
-            String branchName,
-            Mono<List<NewPage>> existingPages,
-            ImportApplicationPermissionProvider permissionProvider) {
-
-        Map<String, String> oldToNewLayoutIds = new HashMap<>();
-        pages.forEach(newPage -> {
-            newPage.setApplicationId(application.getId());
-            if (newPage.getUnpublishedPage() != null) {
-                applicationPageService.generateAndSetPagePolicies(application, newPage.getUnpublishedPage());
-                newPage.setPolicies(newPage.getUnpublishedPage().getPolicies());
-                newPage.getUnpublishedPage().getLayouts().forEach(layout -> {
-                    String layoutId = new ObjectId().toString();
-                    oldToNewLayoutIds.put(layout.getId(), layoutId);
-                    layout.setId(layoutId);
-                });
-            }
-
-            if (newPage.getPublishedPage() != null) {
-                applicationPageService.generateAndSetPagePolicies(application, newPage.getPublishedPage());
-                newPage.getPublishedPage().getLayouts().forEach(layout -> {
-                    String layoutId = oldToNewLayoutIds.containsKey(layout.getId())
-                            ? oldToNewLayoutIds.get(layout.getId())
-                            : new ObjectId().toString();
-                    layout.setId(layoutId);
-                });
-            }
-        });
-
-        return existingPages
-                .flatMapMany(existingSavedPages -> {
-                    Map<String, NewPage> savedPagesGitIdToPageMap = new HashMap<>();
-
-                    existingSavedPages.stream()
-                            .filter(newPage -> !StringUtils.isEmpty(newPage.getGitSyncId()))
-                            .forEach(newPage -> savedPagesGitIdToPageMap.put(newPage.getGitSyncId(), newPage));
-
-                    return Flux.fromIterable(pages).flatMap(newPage -> {
-                        log.debug(
-                                "Importing page: {}",
-                                newPage.getUnpublishedPage().getName());
-                        // Check if the page has gitSyncId and if it's already in DB
-                        if (newPage.getGitSyncId() != null
-                                && savedPagesGitIdToPageMap.containsKey(newPage.getGitSyncId())) {
-                            // Since the resource is already present in DB, just update resource
-                            NewPage existingPage = savedPagesGitIdToPageMap.get(newPage.getGitSyncId());
-                            if (!permissionProvider.hasEditPermission(existingPage)) {
-                                log.error(
-                                        "User does not have permission to edit page with id: {}", existingPage.getId());
-                                return Mono.error(new AppsmithException(
-                                        AppsmithError.ACL_NO_RESOURCE_FOUND, FieldName.PAGE, existingPage.getId()));
-                            }
-                            Set<Policy> existingPagePolicy = existingPage.getPolicies();
-                            copyNestedNonNullProperties(newPage, existingPage);
-                            // Update branchName
-                            existingPage.getDefaultResources().setBranchName(branchName);
-                            // Recover the deleted state present in DB from imported page
-                            existingPage
-                                    .getUnpublishedPage()
-                                    .setDeletedAt(newPage.getUnpublishedPage().getDeletedAt());
-                            existingPage.setDeletedAt(newPage.getDeletedAt());
-                            existingPage.setDeleted(newPage.getDeleted());
-                            existingPage.setPolicies(existingPagePolicy);
-                            return newPageService.save(existingPage);
-                        } else {
-                            // check if user has permission to add new page to the application
-                            if (!permissionProvider.canCreatePage(application)) {
-                                log.error(
-                                        "User does not have permission to create page in application with id: {}",
-                                        application.getId());
-                                return Mono.error(new AppsmithException(
-                                        AppsmithError.ACL_NO_RESOURCE_FOUND,
-                                        FieldName.APPLICATION,
-                                        application.getId()));
-                            }
-                            if (application.getGitApplicationMetadata() != null) {
-                                final String defaultApplicationId =
-                                        application.getGitApplicationMetadata().getDefaultApplicationId();
-                                return newPageService
-                                        .findByGitSyncIdAndDefaultApplicationId(
-                                                defaultApplicationId, newPage.getGitSyncId(), Optional.empty())
-                                        .switchIfEmpty(Mono.defer(() -> {
-                                            // This is the first page we are saving with given gitSyncId in this
-                                            // instance
-                                            DefaultResources defaultResources = new DefaultResources();
-                                            defaultResources.setApplicationId(defaultApplicationId);
-                                            defaultResources.setBranchName(branchName);
-                                            newPage.setDefaultResources(defaultResources);
-                                            return saveNewPageAndUpdateDefaultResources(newPage, branchName);
-                                        }))
-                                        .flatMap(branchedPage -> {
-                                            DefaultResources defaultResources = branchedPage.getDefaultResources();
-                                            // Create new page but keep defaultApplicationId and defaultPageId same for
-                                            // both the
-                                            // pages
-                                            defaultResources.setBranchName(branchName);
-                                            newPage.setDefaultResources(defaultResources);
-                                            newPage.getUnpublishedPage()
-                                                    .setDeletedAt(branchedPage
-                                                            .getUnpublishedPage()
-                                                            .getDeletedAt());
-                                            newPage.setDeletedAt(branchedPage.getDeletedAt());
-                                            newPage.setDeleted(branchedPage.getDeleted());
-                                            // Set policies from existing branch object
-                                            newPage.setPolicies(branchedPage.getPolicies());
-                                            return newPageService.save(newPage);
-                                        });
-                            }
-                            return saveNewPageAndUpdateDefaultResources(newPage, branchName);
-                        }
-                    });
-                })
-                .onErrorResume(error -> {
-                    log.error("Error importing page", error);
-                    return Mono.error(error);
-                });
-    }
-
-    private Mono<NewPage> saveNewPageAndUpdateDefaultResources(NewPage newPage, String branchName) {
-        NewPage update = new NewPage();
-        return newPageService.save(newPage).flatMap(page -> {
-            update.setDefaultResources(
-                    DefaultResourcesUtils.createDefaultIdsOrUpdateWithGivenResourceIds(page, branchName)
-                            .getDefaultResources());
-            return newPageService.update(page.getId(), update);
-        });
-    }
-
-    private Set<String> getLayoutOnLoadActionsForPage(
-            NewPage page,
-            Map<String, String> actionIdMap,
-            Map<String, List<String>> unpublishedActionIdToCollectionIdsMap,
-            Map<String, List<String>> publishedActionIdToCollectionIdsMap) {
-        Set<String> layoutOnLoadActions = new HashSet<>();
-        if (page.getUnpublishedPage().getLayouts() != null) {
-
-            page.getUnpublishedPage().getLayouts().forEach(layout -> {
-                if (layout.getLayoutOnLoadActions() != null) {
-                    layout.getLayoutOnLoadActions()
-                            .forEach(onLoadAction -> onLoadAction.forEach(actionDTO -> {
-                                actionDTO.setId(actionIdMap.get(actionDTO.getId()));
-                                if (!CollectionUtils.sizeIsEmpty(unpublishedActionIdToCollectionIdsMap)
-                                        && !CollectionUtils.isEmpty(
-                                                unpublishedActionIdToCollectionIdsMap.get(actionDTO.getId()))) {
-                                    actionDTO.setCollectionId(unpublishedActionIdToCollectionIdsMap
-                                            .get(actionDTO.getId())
-                                            .get(0));
-                                }
-                                layoutOnLoadActions.add(actionDTO.getId());
-                            }));
-                }
-            });
-        }
-
-        if (page.getPublishedPage() != null && page.getPublishedPage().getLayouts() != null) {
-
-            page.getPublishedPage().getLayouts().forEach(layout -> {
-                if (layout.getLayoutOnLoadActions() != null) {
-                    layout.getLayoutOnLoadActions()
-                            .forEach(onLoadAction -> onLoadAction.forEach(actionDTO -> {
-                                actionDTO.setId(actionIdMap.get(actionDTO.getId()));
-                                if (!CollectionUtils.sizeIsEmpty(publishedActionIdToCollectionIdsMap)
-                                        && !CollectionUtils.isEmpty(
-                                                publishedActionIdToCollectionIdsMap.get(actionDTO.getId()))) {
-                                    actionDTO.setCollectionId(publishedActionIdToCollectionIdsMap
-                                            .get(actionDTO.getId())
-                                            .get(0));
-                                }
-                                layoutOnLoadActions.add(actionDTO.getId());
-                            }));
-                }
-            });
-        }
-
-        layoutOnLoadActions.remove(null);
-        return layoutOnLoadActions;
-    }
-
-    // This method will update the action id in saved page for layoutOnLoadAction
-    private Mono<NewPage> mapActionAndCollectionIdWithPageLayout(
-            NewPage newPage,
-            Map<String, String> actionIdMap,
-            Map<String, List<String>> unpublishedActionIdToCollectionIdsMap,
-            Map<String, List<String>> publishedActionIdToCollectionIdsMap) {
-
-        return Mono.just(newPage)
-                .flatMap(page -> {
-                    return newActionService
-                            .findAllById(getLayoutOnLoadActionsForPage(
-                                    page,
-                                    actionIdMap,
-                                    unpublishedActionIdToCollectionIdsMap,
-                                    publishedActionIdToCollectionIdsMap))
-                            .map(newAction -> {
-                                final String defaultActionId =
-                                        newAction.getDefaultResources().getActionId();
-                                if (page.getUnpublishedPage().getLayouts() != null) {
-                                    final String defaultCollectionId = newAction
-                                            .getUnpublishedAction()
-                                            .getDefaultResources()
-                                            .getCollectionId();
-                                    page.getUnpublishedPage().getLayouts().forEach(layout -> {
-                                        if (layout.getLayoutOnLoadActions() != null) {
-                                            layout.getLayoutOnLoadActions()
-                                                    .forEach(onLoadAction -> onLoadAction.stream()
-                                                            .filter(actionDTO -> StringUtils.equals(
-                                                                    actionDTO.getId(), newAction.getId()))
-                                                            .forEach(actionDTO -> {
-                                                                actionDTO.setDefaultActionId(defaultActionId);
-                                                                actionDTO.setDefaultCollectionId(defaultCollectionId);
-                                                            }));
-                                        }
-                                    });
-                                }
-
-                                if (page.getPublishedPage() != null
-                                        && page.getPublishedPage().getLayouts() != null) {
-                                    page.getPublishedPage().getLayouts().forEach(layout -> {
-                                        if (layout.getLayoutOnLoadActions() != null) {
-                                            layout.getLayoutOnLoadActions()
-                                                    .forEach(onLoadAction -> onLoadAction.stream()
-                                                            .filter(actionDTO -> StringUtils.equals(
-                                                                    actionDTO.getId(), newAction.getId()))
-                                                            .forEach(actionDTO -> {
-                                                                actionDTO.setDefaultActionId(defaultActionId);
-                                                                if (newAction.getPublishedAction() != null
-                                                                        && newAction
-                                                                                        .getPublishedAction()
-                                                                                        .getDefaultResources()
-                                                                                != null) {
-                                                                    actionDTO.setDefaultCollectionId(newAction
-                                                                            .getPublishedAction()
-                                                                            .getDefaultResources()
-                                                                            .getCollectionId());
-                                                                }
-                                                            }));
-                                        }
-                                    });
-                                }
-                                return newAction;
-                            })
-                            .collectList()
-                            .thenReturn(page);
-                })
-                .onErrorResume(error -> {
-                    log.error("Error while updating action collection id in page layout", error);
-                    return Mono.error(error);
-                });
-    }
-
-    /**
-     * This will check if the datasource is already present in the workspace and create a new one if unable to find one
-     *
-     * @param existingDatasources already present datasource in the workspace
-     * @param datasourceStorage   which will be checked against existing datasources
-     * @param workspace           workspace where duplicate datasource should be checked
-     * @return already present or brand new datasource depending upon the equality check
-     */
-    private Mono<Datasource> createUniqueDatasourceIfNotPresent(
-            Flux<Datasource> existingDatasources,
-            DatasourceStorage datasourceStorage,
-            Workspace workspace,
-            String environmentId,
-            ImportApplicationPermissionProvider permissionProvider) {
-        /*
-           1. If same datasource is present return
-           2. If unable to find the datasource create a new datasource with unique name and return
-        */
-        final DatasourceConfiguration datasourceConfig = datasourceStorage.getDatasourceConfiguration();
-        AuthenticationResponse authResponse = new AuthenticationResponse();
-        if (datasourceConfig != null && datasourceConfig.getAuthentication() != null) {
-            copyNestedNonNullProperties(datasourceConfig.getAuthentication().getAuthenticationResponse(), authResponse);
-            datasourceConfig.getAuthentication().setAuthenticationResponse(null);
-            datasourceConfig.getAuthentication().setAuthenticationType(null);
-        }
-
-        return existingDatasources
-                // For git import exclude datasource configuration
-                .filter(ds -> ds.getName().equals(datasourceStorage.getName())
-                        && datasourceStorage.getPluginId().equals(ds.getPluginId()))
-                .next() // Get the first matching datasource, we don't need more than one here.
-                .switchIfEmpty(Mono.defer(() -> {
-                    // check if user has permission to create datasource
-                    if (!permissionProvider.canCreateDatasource(workspace)) {
-                        log.error(
-                                "Unauthorized to create datasource: {} in workspace: {}",
-                                datasourceStorage.getName(),
-                                workspace.getName());
-                        return Mono.error(new AppsmithException(
-                                AppsmithError.ACL_NO_RESOURCE_FOUND,
-                                FieldName.DATASOURCE,
-                                datasourceStorage.getName()));
-                    }
-
-                    if (datasourceConfig != null && datasourceConfig.getAuthentication() != null) {
-                        datasourceConfig.getAuthentication().setAuthenticationResponse(authResponse);
-                    }
-                    // No matching existing datasource found, so create a new one.
-                    datasourceStorage.setIsConfigured(
-                            datasourceConfig != null && datasourceConfig.getAuthentication() != null);
-                    datasourceStorage.setEnvironmentId(environmentId);
-
-                    return datasourceService
-                            .findByNameAndWorkspaceId(datasourceStorage.getName(), workspace.getId(), Optional.empty())
-                            .flatMap(duplicateNameDatasource ->
-                                    getUniqueSuffixForDuplicateNameEntity(duplicateNameDatasource, workspace.getId()))
-                            .map(dsName -> {
-                                datasourceStorage.setName(datasourceStorage.getName() + dsName);
-                                return datasourceService.createDatasourceFromDatasourceStorage(datasourceStorage);
-                            })
-                            .switchIfEmpty(Mono.just(
-                                    datasourceService.createDatasourceFromDatasourceStorage(datasourceStorage)))
-                            .flatMap(datasourceService::createWithoutPermissions);
-                }))
-                .onErrorResume(throwable -> {
-                    log.error("failed to import datasource", throwable);
-                    return Mono.error(throwable);
-                });
-    }
-
-    /**
-     * Here we will be rehydrating the sensitive fields like password, secrets etc. in datasourceStorage while importing the application
-     *
-     * @param datasourceStorage for which sensitive fields should be rehydrated
-     * @param decryptedFields   sensitive fields
-     * @return updated datasourceStorage with rehydrated sensitive fields
-     */
-    private DatasourceStorage updateAuthenticationDTO(
-            DatasourceStorage datasourceStorage, DecryptedSensitiveFields decryptedFields) {
-
-        final DatasourceConfiguration dsConfig = datasourceStorage.getDatasourceConfiguration();
-        String authType = decryptedFields.getAuthType();
-        if (dsConfig == null || authType == null) {
-            return datasourceStorage;
-        }
-
-        if (StringUtils.equals(authType, DBAuth.class.getName())) {
-            final DBAuth dbAuth = decryptedFields.getDbAuth();
-            dbAuth.setPassword(decryptedFields.getPassword());
-            datasourceStorage.getDatasourceConfiguration().setAuthentication(dbAuth);
-        } else if (StringUtils.equals(authType, BasicAuth.class.getName())) {
-            final BasicAuth basicAuth = decryptedFields.getBasicAuth();
-            basicAuth.setPassword(decryptedFields.getPassword());
-            datasourceStorage.getDatasourceConfiguration().setAuthentication(basicAuth);
-        } else if (StringUtils.equals(authType, OAuth2.class.getName())) {
-            OAuth2 auth2 = decryptedFields.getOpenAuth2();
-            AuthenticationResponse authResponse = new AuthenticationResponse();
-            auth2.setClientSecret(decryptedFields.getPassword());
-            authResponse.setToken(decryptedFields.getToken());
-            authResponse.setRefreshToken(decryptedFields.getRefreshToken());
-            authResponse.setTokenResponse(decryptedFields.getTokenResponse());
-            authResponse.setExpiresAt(Instant.now());
-            auth2.setAuthenticationResponse(authResponse);
-            datasourceStorage.getDatasourceConfiguration().setAuthentication(auth2);
-        } else if (StringUtils.equals(authType, BearerTokenAuth.class.getName())) {
-            BearerTokenAuth auth = new BearerTokenAuth();
-            auth.setBearerToken(decryptedFields.getBearerTokenAuth().getBearerToken());
-            datasourceStorage.getDatasourceConfiguration().setAuthentication(auth);
-        }
-        return datasourceStorage;
-    }
-
-    private Mono<Application> importThemes(
-            Application application, ApplicationJson importedApplicationJson, boolean appendToApp) {
-        if (appendToApp) {
-            // appending to existing app, theme should not change
-            return Mono.just(application);
-        }
-        return themeService.importThemesToApplication(application, importedApplicationJson);
-    }
-
-    public Mono<List<Datasource>> findDatasourceByApplicationId(String applicationId, String workspaceId) {
-        // TODO: Investigate further why datasourcePermission.getReadPermission() is not being used.
-        Mono<List<Datasource>> listMono = datasourceService
-                .getAllByWorkspaceIdWithStorages(workspaceId, Optional.empty())
-                .collectList();
-        return newActionService
-                .findAllByApplicationIdAndViewMode(applicationId, false, Optional.empty(), Optional.empty())
-                .collectList()
-                .zipWith(listMono)
-                .flatMap(objects -> {
-                    List<Datasource> datasourceList = objects.getT2();
-                    List<NewAction> actionList = objects.getT1();
-                    List<String> usedDatasource = actionList.stream()
-                            .map(newAction -> newAction
-                                    .getUnpublishedAction()
-                                    .getDatasource()
-                                    .getId())
-                            .collect(Collectors.toList());
-
-                    datasourceList.removeIf(datasource -> !usedDatasource.contains(datasource.getId()));
-
-                    return Mono.just(datasourceList);
-                });
     }
 
     @Override
@@ -1905,6 +670,32 @@ public class ImportApplicationServiceCEImpl implements ImportApplicationServiceC
                         applicationImportDTO.setIsPartialImport(false);
                     }
                     return applicationImportDTO;
+                });
+    }
+
+    @Override
+    public Mono<List<Datasource>> findDatasourceByApplicationId(String applicationId, String workspaceId) {
+        // TODO: Investigate further why datasourcePermission.getReadPermission() is not being used.
+        Mono<List<Datasource>> listMono = datasourceService
+                .getAllByWorkspaceIdWithStorages(workspaceId, Optional.empty())
+                .collectList();
+        return newActionService
+                .findAllByApplicationIdAndViewMode(applicationId, false, Optional.empty(), Optional.empty())
+                .collectList()
+                .zipWith(listMono)
+                .flatMap(objects -> {
+                    List<Datasource> datasourceList = objects.getT2();
+                    List<NewAction> actionList = objects.getT1();
+                    List<String> usedDatasource = actionList.stream()
+                            .map(newAction -> newAction
+                                    .getUnpublishedAction()
+                                    .getDatasource()
+                                    .getId())
+                            .collect(Collectors.toList());
+
+                    datasourceList.removeIf(datasource -> !usedDatasource.contains(datasource.getId()));
+
+                    return Mono.just(datasourceList);
                 });
     }
 
@@ -1996,35 +787,6 @@ public class ImportApplicationServiceCEImpl implements ImportApplicationServiceC
             return importApplicationInWorkspace(
                     workspaceId, applicationJson, applicationId, branchName, true, permissionProvider);
         });
-    }
-
-    private Map<String, String> updateNewPagesBeforeMerge(List<NewPage> existingPages, List<NewPage> newPagesList) {
-        Map<String, String> newToOldToPageNameMap = new HashMap<>(); // maps new names with old names
-
-        // get a list of unpublished page names that already exists
-        List<String> unpublishedPageNames = existingPages.stream()
-                .filter(newPage -> newPage.getUnpublishedPage() != null)
-                .map(newPage -> newPage.getUnpublishedPage().getName())
-                .collect(Collectors.toList());
-
-        // modify each new page
-        for (NewPage newPage : newPagesList) {
-            newPage.setPublishedPage(null); // we'll not merge published pages so removing this
-
-            // let's check if page name conflicts, rename in that case
-            String oldPageName = newPage.getUnpublishedPage().getName(),
-                    newPageName = newPage.getUnpublishedPage().getName();
-
-            int i = 1;
-            while (unpublishedPageNames.contains(newPageName)) {
-                i++;
-                newPageName = oldPageName + i;
-            }
-            newPage.getUnpublishedPage().setName(newPageName); // set new name. may be same as before or not
-            newPage.getUnpublishedPage().setSlug(TextUtils.makeSlug(newPageName)); // set the slug also
-            newToOldToPageNameMap.put(newPageName, oldPageName); // map: new name -> old name
-        }
-        return newToOldToPageNameMap;
     }
 
     /**
