+++ app/server/appsmith-server/src/main/java/com/appsmith/server/imports/internal/ImportApplicationServiceImpl.java
@@ -1,16 +1,19 @@
 package com.appsmith.server.imports.internal;
 
-import com.appsmith.server.actioncollections.base.ActionCollectionService;
+import com.appsmith.external.models.Datasource;
 import com.appsmith.server.datasources.base.DatasourceService;
-import com.appsmith.server.jslibs.base.CustomJSLibService;
+import com.appsmith.server.domains.ActionCollection;
+import com.appsmith.server.domains.CustomJSLib;
+import com.appsmith.server.domains.NewAction;
+import com.appsmith.server.domains.NewPage;
+import com.appsmith.server.domains.Plugin;
+import com.appsmith.server.domains.Theme;
+import com.appsmith.server.imports.importable.ImportableService;
 import com.appsmith.server.newactions.base.NewActionService;
-import com.appsmith.server.newpages.base.NewPageService;
 import com.appsmith.server.repositories.PermissionGroupRepository;
-import com.appsmith.server.repositories.PluginRepository;
 import com.appsmith.server.services.AnalyticsService;
 import com.appsmith.server.services.ApplicationPageService;
 import com.appsmith.server.services.ApplicationService;
-import com.appsmith.server.services.SequenceService;
 import com.appsmith.server.services.SessionUserService;
 import com.appsmith.server.services.WorkspaceService;
 import com.appsmith.server.solutions.ActionPermission;
@@ -18,32 +21,25 @@ import com.appsmith.server.solutions.ApplicationPermission;
 import com.appsmith.server.solutions.DatasourcePermission;
 import com.appsmith.server.solutions.PagePermission;
 import com.appsmith.server.solutions.WorkspacePermission;
-import com.appsmith.server.themes.base.ThemeService;
 import com.google.gson.Gson;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Primary;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.reactive.TransactionalOperator;
 
 @Slf4j
-@Component
+@Service
 @Primary
 public class ImportApplicationServiceImpl extends ImportApplicationServiceCEImpl implements ImportApplicationService {
 
     public ImportApplicationServiceImpl(
             DatasourceService datasourceService,
             SessionUserService sessionUserService,
-            PluginRepository pluginRepository,
             WorkspaceService workspaceService,
             ApplicationService applicationService,
-            NewPageService newPageService,
             ApplicationPageService applicationPageService,
             NewActionService newActionService,
-            SequenceService sequenceService,
-            ActionCollectionService actionCollectionService,
-            ThemeService themeService,
             AnalyticsService analyticsService,
-            CustomJSLibService customJSLibService,
             DatasourcePermission datasourcePermission,
             WorkspacePermission workspacePermission,
             ApplicationPermission applicationPermission,
@@ -51,21 +47,22 @@ public class ImportApplicationServiceImpl extends ImportApplicationServiceCEImpl
             ActionPermission actionPermission,
             Gson gson,
             TransactionalOperator transactionalOperator,
-            PermissionGroupRepository permissionGroupRepository) {
+            PermissionGroupRepository permissionGroupRepository,
+            ImportableService<Plugin> pluginImportableService,
+            ImportableService<Theme> themeImportableService,
+            ImportableService<NewPage> newPageImportableService,
+            ImportableService<CustomJSLib> customJSLibImportableService,
+            ImportableService<Datasource> datasourceImportableService,
+            ImportableService<NewAction> newActionImportableService,
+            ImportableService<ActionCollection> actionCollectionImportableService) {
         super(
                 datasourceService,
                 sessionUserService,
-                pluginRepository,
                 workspaceService,
                 applicationService,
-                newPageService,
                 applicationPageService,
                 newActionService,
-                sequenceService,
-                actionCollectionService,
-                themeService,
                 analyticsService,
-                customJSLibService,
                 datasourcePermission,
                 workspacePermission,
                 applicationPermission,
@@ -73,6 +70,13 @@ public class ImportApplicationServiceImpl extends ImportApplicationServiceCEImpl
                 actionPermission,
                 gson,
                 transactionalOperator,
-                permissionGroupRepository);
+                permissionGroupRepository,
+                pluginImportableService,
+                themeImportableService,
+                newPageImportableService,
+                customJSLibImportableService,
+                datasourceImportableService,
+                newActionImportableService,
+                actionCollectionImportableService);
     }
 }
