+++ app/server/appsmith-server/src/main/java/com/appsmith/server/jslibs/imports/CustomJSLibImportableServiceImpl.java
@@ -1,0 +1,14 @@
+package com.appsmith.server.jslibs.imports;
+
+import com.appsmith.server.domains.CustomJSLib;
+import com.appsmith.server.imports.importable.ImportableService;
+import com.appsmith.server.jslibs.base.CustomJSLibService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CustomJSLibImportableServiceImpl extends CustomJSLibImportableServiceCEImpl
+        implements ImportableService<CustomJSLib> {
+    public CustomJSLibImportableServiceImpl(CustomJSLibService customJSLibService) {
+        super(customJSLibService);
+    }
+}
