+++ app/server/appsmith-server/src/main/java/com/appsmith/server/plugins/imports/PluginImportableServiceImpl.java
@@ -1,0 +1,13 @@
+package com.appsmith.server.plugins.imports;
+
+import com.appsmith.server.domains.Plugin;
+import com.appsmith.server.imports.importable.ImportableService;
+import com.appsmith.server.plugins.base.PluginService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class PluginImportableServiceImpl extends PluginImportableServiceCEImpl implements ImportableService<Plugin> {
+    public PluginImportableServiceImpl(PluginService pluginService) {
+        super(pluginService);
+    }
+}
